+++ src/metrics/flowsom_mapping_similarity/script.R
@@ -119,5 +119,4 @@ output <- anndata::AnnData(
 
 cat("Output AnnData object:\n")
 print(output)
-
 output$write_h5ad(par[["output"]], compression = "gzip")